% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Imports.R
\docType{package}
\name{MT4113GroupProject-package}
\alias{MT4113GroupProject}
\alias{MT4113GroupProject-package}
\title{MT4113GroupProject: Unified Optimiser}
\description{
This package is designed to optimise a wide variety of functions via one of five methods—grid search, bisection method, univariate Newton, multivariate Newton and Gauss-Newton. Users have the option to select one of these methods, and also input a variety of analytic functions to speed up processing time if desired. All functions in this package are tied to the parent function 'funoptim', and this should be the only function users need to call.
}
\keyword{internal}
