% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVN.R
\name{Norm_Ratio}
\alias{Norm_Ratio}
\title{Norm Ratio}
\usage{
Norm_Ratio(delta, theta)
}
\arguments{
\item{delta}{step value for each parameter}

\item{theta}{estimate of the minimised values}
}
\value{
the ratio of the Euclidean norm of the inputted vectors
}
\description{
Computes the ratio of the Euclidean norm of 2 vectors
}
\author{
Holly Goldsmith
}
