% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS.R
\name{BS}
\alias{BS}
\title{Bisection Method}
\usage{
BS(
  f,
  inits,
  data = NULL,
  minimum = TRUE,
  tol = 1e-08,
  maxit = 100,
  method = BS,
  gradfn = NULL,
  hessfn = NULL,
  jacobfn = NULL
)
}
\arguments{
\item{f}{Function to be optimised}

\item{inits}{A length-2 numeric vector giving the initial search interval.
The gradient at these two points must have opposite signs.}

\item{data}{(optional) Dataframe passed to the function if needed.}

\item{minimum}{(optional) Search for minimum or maximum (takes TRUE if minimum search, FALSE if maximum search)}

\item{tol}{(optional) Tolerance level}

\item{maxit}{(optional) Maximum number of iterations run before stopping}

\item{method}{(optional) NULL variable, included to match parent function inputs}

\item{gradfn}{(optional) NULL variable, included to match parent function inputs}

\item{hessfn}{(optional) NULL variable, included to match parent function inputs}

\item{jacobfn}{(optional) NULL variable, included to match parent function inputs}
}
\value{
A list containing:
\item{estimate}{Optimised estimate}
\item{feval}{Function evaluated at optimised estimate}
\item{grad}{Gradient of function at optimised estimate}
\item{tolerance}{Final interval width}
\item{conv}{Whether or not the optimisation converged. 0 - converged, 1 - failed, 2 - max iterations reached}
\item{niter}{Number of iterations}
}
\description{
The Bisection method works by taking two initial points, finding their derivatives and then if they
are opposite signs we look at the midpoint, repeating until we find where the derivative is zero
or within tolerance.
This method requires a smooth function.
}
\examples{
test_f<-function(x){
return (x^5-3*x^4+12*x^3-5*x^2-3)
}
test_inits<-c(0.1,5)
BS(test_f,test_inits)

}
\references{
Swallow, B. (2025). Univariate Optimisation.
University of St Andrews.
\url{https://moody.st-andrews.ac.uk/moodle/pluginfile.php/2128840/mod_resource/content/4/_book/univariate-optimization.html#bisection-method}
}
\author{
Charles Beer
}
