% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GN.R
\name{GN}
\alias{GN}
\title{Gauss-Newton Method}
\usage{
GN(
  f,
  inits,
  data = NULL,
  minimum = TRUE,
  tol = 1e-10,
  maxit = 1000,
  method = "GN",
  gradfn = NULL,
  hessfn = NULL,
  jacobfn = NULL
)
}
\arguments{
\item{f}{the function to be optimised}

\item{inits}{a vector of initial guesses for each parameter to be optimised.}

\item{data}{dataframe containing the data}

\item{minimum}{not relevant, included to match parent function inputs}

\item{tol}{tolerance level}

\item{maxit}{maximum number of iterations run before stopping}

\item{method}{identifier, takes "GN" only to allow the parent function to call on it}

\item{gradfn}{NULL variable, included to match parent function inputs}

\item{hessfn}{NULL variable, included to match parent function inputs}

\item{jacobfn}{(optional) Jacobian function of residuals, uses numerical approximation otherwise}
}
\value{
A list containing:
\item{estimate}{Optimised estimate}
\item{feval}{Function evaluated at optimised estimate}
\item{grad}{Gradient of function at optimised estimate}
\item{tolerance}{Tolerance level reached through optimisation}
\item{conv}{Whether or not the optimisation converged. 0 - converged, 1 - failed, 2 - max iterations reached}
\item{niter}{Number of iterations ran}
}
\description{
Performs non-linear least squares optimisation using Gauss-Newton method
}
\examples{
f <- function(theta, data){
  a <- theta[1]
  b <- theta[2]
  return(a * exp(b * data$x))
}

data <- list(x = 1:10, y = 2.5 * exp(-1.5 * 1:10))
inits <- c(1, -1)

GN(f, inits, data)
}
\references{
Swallow, B. (2025). Non-Linear Least Squares.
University of St Andrews.
\url{https://moody.st-andrews.ac.uk/moodle/pluginfile.php/2128840/mod_resource/content/4/_book/non-linear-least-squares.html#gauss-newton-method}
}
\author{
Ilana Goldman
}
