% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS.R
\name{GS}
\alias{GS}
\title{Grid search Method}
\usage{
GS(
  f,
  inits,
  data = NULL,
  minimum = TRUE,
  tol = 1e-08,
  maxit = 100,
  method = NULL,
  gradfn = NULL,
  hessfn = NULL,
  jacobfn = NULL
)
}
\arguments{
\item{f}{Function to be optimised}

\item{inits}{An initial guess, the grid will be on +- 10 of this point}

\item{data}{(optional) Dataframe passed to the function if needed.}

\item{minimum}{(optional) Search for minimum or maximum (takes TRUE if minimum search, FALSE if maximum search)}

\item{tol}{(optional) Tolerance level}

\item{maxit}{(optional) Maximum number of iterations run before stopping}

\item{method}{(optional) NULL variable, included to match parent function inputs}

\item{gradfn}{(optional) NULL variable, included to match parent function inputs}

\item{hessfn}{(optional) NULL variable, included to match parent function inputs}

\item{jacobfn}{(optional) NULL variable, included to match parent function inputs}
}
\value{
A list containing:
\item{estimate}{Optimised estimate}
\item{feval}{Function evaluated at optimised estimate}
\item{grad}{Gradient of function at optimised estimate}
\item{tolerance}{Tolerance level reached through optimisation}
\item{conv}{Whether or not the optimisation converged. 0 - converged, 1 - failed, 2 - max iterations reached}
\item{niter}{Number of iterations}
}
\description{
The grid search method works by evaluating a function on points of a grid and picking
the point which has either the maximum or minimum value depending on the given criteria
}
\examples{
test_f<-function(x){
return ((x-0.25)*(x-0.6)*(x-5)*(x-3))
}
test_inits<-c(0)
GS(test_f,test_inits,minimum=TRUE)

}
\references{
Swallow, B. (2025). Univariate Optimization.
University of St Andrews.
\url{https://moody.st-andrews.ac.uk/moodle/pluginfile.php/2128840/mod_resource/content/4/_book/univariate-optimization.html#grid-search}
}
\author{
Charles Beer
}
