% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UVN.R
\name{UVN}
\alias{UVN}
\title{Univariate Newton Optimisation}
\usage{
UVN(
  f,
  inits,
  data,
  minimum,
  tol,
  maxit,
  method = "UVN",
  gradfn,
  hessfn,
  jacobfn
)
}
\arguments{
\item{f}{Function to be minimised}

\item{inits}{The initial value; a numeric scalar giving the starting point}

\item{data}{(Optional) Single column dataframe of data in relation to the parameter to be optimised}

\item{minimum}{Search for minimum or maximum (takes TRUE if minimum search, FALSE if maximum search)}

\item{tol}{Tolerance level}

\item{maxit}{Maximum number of iterations run before stopping}

\item{method}{Identifier, takes "UVN" only to allow the parent function to call on it}

\item{gradfn}{(Optional) Gradient function of f, uses finite differencing otherwise}

\item{hessfn}{(Optional) Hessian function of f, uses finite differencing otherwise}

\item{jacobfn}{NULL variable, included to match parent function inputs}
}
\value{
A list containing:
\item{estimate}{Optimised estimate}
\item{feval}{Function evaluated at optimised estimate}
\item{grad}{Gradient of function at optimised estimate}
\item{tolerance}{Tolerance level reached through optimisation}
\item{conv}{Whether or not the optimisation converged. 0 - converged, 1 - did not converge, 2 - max iterations reached}
\item{niter}{Number of iterations run}
}
\description{
Performs univariate optimisation using Newton's method
}
\examples{
f <- function(theta){
x <- theta[1]
return((x - 33)^3)
}
UVN(f = f, inits = 10, data = NULL, minimum = TRUE,
tol = 1e-6, maxit = 100, method = "UVN",
gradfn = NULL, hessfn = NULL, jacobfn = NULL)

}
\references{
Swallow, B. (2025). Univariate Optimisation.
University of St Andrews.
\url{https://moody.st-andrews.ac.uk/moodle/pluginfile.php/2128840/mod_resource/content/4/_book/univariate-optimization.html#bisection-method}
}
\author{
Chenyu Lin
}
